CHE-84 Thread controller directory restructure
