+++ server/controllers/threadController/index.ts
@@ -1,0 +1,15 @@
+import createThread from './createThread/createThread';
+import listThreadsByForumId from './listThreadsByForumId/listThreadsByForumId';
+import getThreadById from './getThreadById/getThreadById';
+import updateThread from './updateThread/updateThread';
+import deleteThread from './deleteThread/deleteThread';
+import getAllThreads from './getAllThreads/getAllThreads';
+
+export {
+  createThread,
+  listThreadsByForumId,
+  getThreadById,
+  updateThread,
+  deleteThread,
+  getAllThreads,
+};
